+++ src/vmware/HISTORY.md
@@ -7,6 +7,8 @@
 - Add `--management-network` and `--uplink-network` to `az vmware addon hcx create` and `az vmware addon hcx update`
 - Add `az vmware cluster host` command group
 - Add `az vmware datastores pure-storage-volume` command group
+- Add `--zones` to `az vmware private-cloud create`
+- [BREAKING CHANGE] Remove `--nsxt-password` and `--vcenter-password` as parameters for `az vmware private-cloud create` and `az vmware private-cloud update`
 
 ## 7.2.0 (2025-02)